un-PIMPL videometadata
It has been practically unchanged for 15 years, and since it is a data structure, any change to the private variables needed a change to the header, defeating the point of the PIMPL idiom.  Furthermore, it is not included by a large number of files, so the extra indirections are wasteful and unnecessary.

Also cleanup class: merge multiple public sections
include header directly in cpp file
use default destructor, copy constructor, copy assignment