videometadata: default insert date to QDate::currentDate()
instead of explicitly immediately setting it.

Use QDate::currentDate instead of MythDate(Time)::current().date()
since we probably want the local date and time zone corrections are
unnecessary here because we are not using the time.  By only using the date,
which is inherently ambiguous as a period of 24 hours, any time zone correction
is irrelevant.