libmyth/mythcontext: remove gContext
Convert it into an instance so the destructor is called automatically.

mythtv-setup:
SetupMenuCallback() does not use the data pointer, so feed it with nullptr.

mythfrontend:
TVMenuCallback() also does not use the data pointer, so feed it with nullptr.

mythbackend:
httpconfig.cpp: The only HttpConfig is created in MediaServer::Init() which is
only called in run_backend() after the MythContext is created in main(), so that
check was always false.

V2Myth::s_WebOnlyStartup: I would have prefered to avoid global state and pass
it as a parameter to V2Myth() via run_backend() and run_setup_webserver() but
the entire class is already static.

V2Myth is a better place for WebOnlyStartup since that is the only place it is
used.

CleanupGuards:
Move the CleanupGuard instance after the MythContext, so it is destroyed first.
~MythCoreContextPrivate() calls MThread::Cleanup() which wants all threads to
already be destroyed.  Move DestroyMythMainWindow() into ~MythContext() so the
static MythUDP instance destroys its thread.

ReferenceCounter::PrintDebug(): remove extraneous ui include and move into
~MythContext() so it is called after all objects should be destroyed.