Modernize mythtranscode to use std::vector. (pes_in_t)
Convert the pes_in_t "buf" member from a raw memory pointer to a
std::vector.  This delegates the memory cleanup to the standard
library.  This code is never called by MythTV, as far as I can
determine.  It exists solely because the replex code was copied in its
entirety into MythTV.