Modernize lirc_client.cpp to use std::string. (lirc_code)
Convert raw memory pointers in the lirc_code structure to use
std::string.  This allows the code to be simplified by delegating all
the memory management code to the standard library.

This also replaces the hack of defining LIRC_ALL as "((char *) (-1))"
with the string "ALL", since the places it was used are now all
std::strings.