Modernize lirc_client.cpp to use std::string. (lirc_config_entry)
Convert two raw memory pointers in the lirc_config_entry structure to
use std::string.  This allows the code to be simplified by delegating
all the memory management code to the standard library.

Also since the structure is now being allocated with "new", define the
initial values for each member in the structure definition instead of
requiring assign statements after the "malloc".